requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Dashboard
streamlit==1.28.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Téléphonie
twilio==8.10.0

# HTTP
requests==2.31.0
httpx==0.25.0

# LLM
google-generativeai==0.3.0

# Recherche
google-search-results==2.4.2

# Tunneling
pyngrok==7.0.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilitaires
pandas==2.1.3
